Major work to get CLI tests working
pytest fixture to insert sqlite db session into the click context
object. This is required as sqlalchemy sessions with
`sqlite:///:memory:` can end up having multiple separate databases in
memory. So we have to ensure a single session factory is used by both
the tests and the click code under test.

Sort out the pytest fixtures to ensure that tables are
created for each test and dropped after each test, so we don't have data
left over from one test affecting another test.

Consistently use sqlite in-memory databases for the integration tests.

Ensure that SQLAlchemy makes timezone-aware datetime objects when
fetching them from the database.

Add a wrapper type for UUID - so that we use the native postgres UUID
type with postgres and a string in sqlite.

Add `create_detached_copy()` methods on the domain objects - so we can
create a copy of an object from a SQLAlchemy session before existing the
uow block (and so closing the session). If we don't do that, then we
can't access the attributes of the object later.

Move various bits of logic out of the cli code and into services. And
just remove some options from the cli, so we have less code to test.

Allow creating a user in the service layer without an invite code -
sometimes we need to do that.

Stop trying to create invites without a user - we just always have an
admin user we refer to.