Fix bug where invite codes did not get used
Because the function tried to do too much, so had lots of conditionals
to work - and when we used it for real it did not "use" the invite.

Fixed by splitting it into two functions - so the "use" function will
throw an error if it can't actually use the invite.

Also added an integration test to exercise the behaviour at the service
layer, checking the database with a fresh session afterwards