find templates and static when the code is properly installed
in which case the code will be under
`.../lib/python3.x/site-packages/opendlp/...` - and so the trick of
looking for `templates/` and `static/` 3 directories up will not work.

We still default to that, but first check for a `PROJECT_ROOT`
environment variable. If we find it, we just use that.

And we set that in the Dockerfile, which is the place where we do a
proper install. So we don't actually need to set it in the `.env` file.