Get tests working with config now we install project properly
The code to find the static, templates and translations directories
broke because it assumed it was just 3 levels up from flask_app.py -
which is true when running directly, or when using an editable install
(via `uv pip install -e .`)

But the editable install becomes uninstalled a lot, so we've changed to
a proper build and install of the opendlp package. This also works
better with the Dockerfile set up.

But that means we need to find the `<git root>/backend/` directory a bit
more carefully. So now we try (in order):

- explicitly set by the PROJECT_ROOT environment variable
- 3 directories up from `flask_app.py`
- see if we have the `GITHUB_WORKSPACE` environment variable - this is
  for when being run by a github action.
- check the current working directory - for when being run by tests
  locally.

Also update line numbers in the secrets baseline.