Remove restriction on assembly first_assembly_date
We were saying it had to be in the future, but team decided we should be
more flexible than that.

Particularly for updates, we might want to edit late in the day. But
maybe also we're taking on replacements or a later round of an assembly
that started before it was added to OpenDLP.