Show/hide the team-based form elements
On the form to configure the Google Spreadsheet for an assembly, there
are 3 form elements that are automatically set when you choose a
particular team - ID Column, Address Columns and Columns To Keep. If the
"team" element is set to anything other than Custom, those 3 form
elements will be hidden.