Improved wording and fieldsets
After reviewing with the team we're more consistent with talking about
"respondents" rather than "registrants"

We also group the tab name settings in the Google Spreadsheet
Configuration form, to make it clearer to users.

And stop saying "Edit Google Spreadsheet" - which users can think means
opening the spreadsheet itself. When it actually means opening the
configuration.