Ensure we never use stale cached CSS file
by using a query parameter based on the hash of the CSS file.

It is read by a cached function - so we don't re-read every request.
That is then added to the template context by a context processor, and
the base template then adds the hash to the CSS URL.