Test infrastructure can substitute CSV data for google sheets
This is for BDD tests in particular (though it could be used by other
tests).

The idea is that we have a little override from environment variables.
When it is live, instead of trying to talk to Google Spreadsheets via
the API, it will instead just read/write some local CSV files. That
means the code in this repo is fully exercised and selection algorithms
can run fully.