Dynamic search for users (when adding to assembly)
Now you can type and have dynamic results to find the users you want
when adding users to an assembly.  We allow multiple space-separated
fragments to be used in the search box.  The results come back via HTMX
and clicking on them fills in the form with the selected user.