Add opendlp version to footer - date and shorthash of last commit
Find the version of the OpenDLP app. It should be generated by running:

    git show --no-patch --format='%cd %h' --date=format:'%Y-%m-%d' HEAD

which will show the date and short hash of the last commit. For deployment, this will have
been written to the file `generated_version.txt` - so first check for that.

If not, or if that file only contains "HEAD" then check if the parent directory contains ".git"
then try running the git command.

Otherwise return UNKNOWN