Get run_report etc for long-dead selection jobs
By getting it from the SelectionRunRecord if the Celery task result is
no longer available.

Celery will keep track of task results for a while (default 24 hours)
but after that drops them. But we now record the important parts of the
result on the SelectionRunRecord in postgres as well as the result. This
means we can access them after the fact.